requirements.txt
# Data Sources
oandapyV20==0.7.2
yfinance==0.2.32
feedparser==6.0.10

# ML & Data Processing
xgboost==2.0.3
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# NLP - FinBERT
transformers==4.36.0
torch==2.1.1

# Technical Analysis
TA-Lib==0.4.28

# Database

# Scheduling & Monitoring
schedule==1.2.0
python-telegram-bot==20.7

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

